aplicando padrao MVC e refatorando o código
