+++ api/routes/petsRoute.ts
@@ -1,0 +1,8 @@
+import { Router } from "express";
+import { searchAllPets } from "../controllers/pets";
+
+const router = Router();
+
+router.post("/api/pets", searchAllPets);
+
+export default router;
